create group functions

+++ database/migrations/2023_12_06_000003_create_groups_table.php
@@ -13,7 +13,8 @@ return new class extends Migration {
         Schema::create('groups', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->Integer('owner_id');
+            $table->unsignedBigInteger('owner_id');
+
             $table->timestamps();
         });
     }
